requirements.txt
spacy==3.7.4
transformers==4.41.2
datasets==2.20.0
torch==2.3.1
accelerate==0.31.0
seqeval==1.2.2
mlflow==2.14.1
fastapi==0.110.3
uvicorn[standard]==0.30.1
pydantic==2.7.4
sqlalchemy==2.0.31
alembic==1.13.3
psycopg2-binary==2.9.9
pyyaml==6.0.1
python-dotenv==1.0.1
